Move mixin class handling unit conversion to own file
